+++ tests/jvm/src/test/scala/scala/meta/tests/metacp/BaseMetacpSuite.scala
@@ -21,7 +21,7 @@ abstract class BaseMetacpSuite extends BaseCliSuite {
   private val tmp = Files.createTempDirectory("metacp")
   tmp.toFile.deleteOnExit()
 
-  def check(name: String, classpath: () => String): Unit = {
+  def checkMetacp(name: String, classpath: () => String): Unit = {
     test(name) {
       val cp = classpath()
       val args = Array[String](
@@ -37,97 +37,22 @@ abstract class BaseMetacpSuite extends BaseCliSuite {
     }
   }
 
-  def checkClassSignature(name: String, classpath: () => String): Unit = {
-    checkSignatureCallback(name, classpath) { node: ClassNode =>
-      List(
-        (node.signature, { () =>
-          if (node.signature != null) {
-            ClassSignatureSuite.assertRoundtrip(node.signature)
-          }
-        })
-      )
-    }
-  }
-
-  def checkMethodSignature(name: String, classpath: () => String): Unit = {
-    checkSignatureCallback(name, classpath) { node: ClassNode =>
-      import scala.collection.JavaConverters._
-      import scala.tools.asm.tree.MethodNode
-      node.methods.asScala.iterator.map { method: MethodNode =>
-        val signature = if (method.signature == null) method.desc else method.signature
-        (signature, { () =>
-          ClassSignatureSuite.assertMethodRoundtrip(signature)
-        })
-      }.toList
-    }
-  }
-
-  def checkSignatureCallback(name: String, classpath: () => String)(
-      callback: ClassNode => List[(String, () => Unit)]
-  ): Unit = {
-    test(name) {
-      val failingSignatures = ArrayBuffer.empty[String]
-      Classpath(classpath()).visit { root =>
-        new java.nio.file.SimpleFileVisitor[Path] {
-          override def visitFile(file: Path, attrs: BasicFileAttributes): FileVisitResult = {
-            if (PathIO.extension(file) == "class") {
-              val bytes = Files.readAllBytes(file)
-              val node = Javacp.asmNodeFromBytes(bytes)
-              val tests = callback(node)
-              tests.foreach {
-                case (signature, unsafe) =>
-                  try {
-                    unsafe()
-                  } catch {
-                    case NonFatal(e) =>
-                      println(signature)
-                      failingSignatures += signature
-                  }
-              }
-            }
-            FileVisitResult.CONTINUE
-          }
-        }
-      }
-      Files.write(
-        java.nio.file.Paths.get("signatures.txt"),
-        failingSignatures.mkString("\n").getBytes(StandardCharsets.UTF_8)
-      )
-    }
-  }
-
-  case class Coordinates(organization: String, artifact: String, version: String) {
-    def name = List(organization, artifact, version).mkString(File.pathSeparator)
-    def classpath: () => String = { () =>
-      val jars = Jars
-        .fetch(organization, artifact, version)
-        .filterNot(_.toString.contains("scala-lang"))
-      jars.mkString(File.pathSeparator)
-    }
+  def checkLibrary(coordinates: Coordinates): Unit = {
+    checkMetacp(coordinates.name, coordinates.classpath)
   }
 
   val scalameta = Coordinates("org.scalameta", "scalameta_2.12", "3.2.0")
   val akka = Coordinates("com.typesafe.akka", "akka-testkit_2.12", "2.5.9")
   val spark = Coordinates("org.apache.spark", "spark-sql_2.11", "2.2.1")
 
-  def checkLibrary(coordinates: Coordinates): Unit = {
-    check(
-      coordinates.name,
-      coordinates.classpath
-    )
-  }
+}
 
-  def checkClassSignatureLibrary(coordinates: Coordinates): Unit = {
-    checkClassSignature(
-      coordinates.name,
-      coordinates.classpath
-    )
-  }
-
-  def checkMethodSignatureLibrary(coordinates: Coordinates): Unit = {
-    checkMethodSignature(
-      coordinates.name,
-      coordinates.classpath
-    )
+case class Coordinates(organization: String, artifact: String, version: String) {
+  def name = List(organization, artifact, version).mkString(File.pathSeparator)
+  def classpath: () => String = { () =>
+    val jars = Jars
+      .fetch(organization, artifact, version)
+      .filterNot(_.toString.contains("scala-lang"))
+    jars.mkString(File.pathSeparator)
   }
 }
