+++ tests/jvm/src/test/scala/scala/meta/tests/metacp/MetacpSuite.scala
@@ -2,7 +2,7 @@ package scala.meta.tests.metacp
 
 class MetacpSuite extends BaseMetacpSuite {
 
-  check("scala-library", () => scalaLibraryJar)
+  checkMetacp("scala-library", () => scalaLibraryJar)
 
   checkLibrary(scalameta)
   checkLibrary(akka)
